지원 페이지 metadata/server-client 분리
